yolo_based/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
Pillow>=8.0.0

# YOLOv8
ultralytics>=8.0.0

# Tracking
filterpy>=1.4.5
scikit-image>=0.19.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.900 